Fixed installation of initial cross packages
Bugfix:
Introducing DESTDIR installation caused the installation of the initial
cross packages to fail due to conflicts of ownership.
The fix is to install the cross packages directly to $LFS/tools
without using an intermediate DESTDIR step.

Update:
An optional __postconfig function was introduces to try and cure the above
bug. No longer required following the above fix.
Package creation was previously part of the installation function, but has
been moved to the package installation function.